moved unban command
